2.9. Переносит функции для работы с датой в Utils
